backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-dotenv>=0.19.0
pandas>=1.5.0
sqlalchemy>=1.4.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
openai>=2.0.0,<3.0.0
lxml>=4.9.0
numpy>=1.21.0
plotly>=5.15.0
pg8000>=1.29.0
pymysql>=1.0.0
cloud-sql-python-connector[pg8000]>=1.0.0
google-auth>=2.0.0
python-multipart>=0.0.6
requests>=2.31.0
